third_party/blink/web_tests/platform/generic/animations/interpolation/display-interpolation-expected.txt

This is a testharness.js-based test.
PASS This test uses interpolation-test.js.
PASS CSS Animations: property <display> from [none] to [flex] at (-0.3) is [block]
PASS CSS Animations: property <display> from [none] to [flex] at (0) is [block]
PASS CSS Animations: property <display> from [none] to [flex] at (0.3) is [block]
PASS CSS Animations: property <display> from [none] to [flex] at (0.6) is [block]
PASS CSS Animations: property <display> from [none] to [flex] at (1) is [block]
PASS CSS Animations: property <display> from [none] to [flex] at (1.5) is [block]
FAIL Web Animations: property <display> from [none] to [flex] at (-0.3) is [block] assert_equals: expected "none " but got "block "
FAIL Web Animations: property <display> from [none] to [flex] at (0) is [block] assert_equals: expected "none " but got "block "
FAIL Web Animations: property <display> from [none] to [flex] at (0.3) is [block] assert_equals: expected "none " but got "block "
FAIL Web Animations: property <display> from [none] to [flex] at (0.5) is [block] assert_equals: expected "flex " but got "block "
FAIL Web Animations: property <display> from [none] to [flex] at (0.6) is [block] assert_equals: expected "flex " but got "block "
FAIL Web Animations: property <display> from [none] to [flex] at (1) is [block] assert_equals: expected "flex " but got "block "
FAIL Web Animations: property <display> from [none] to [flex] at (1.5) is [block] assert_equals: expected "flex " but got "block "
Harness: the test ran to completion.
